과제 완성

+++ springs/meetingex/src/main/java/com/example/meetingex/service/MeetingService.java
@@ -3,19 +3,23 @@ package com.example.meetingex.service;
 
 import com.example.meetingex.dto.ErrorResponseDto;
 import com.example.meetingex.dto.MeetingDto;
+import com.example.meetingex.dto.ParticipantsDto;
 import com.example.meetingex.entity.Meeting;
+import com.example.meetingex.entity.Schedule;
+import com.example.meetingex.entity.User;
 import com.example.meetingex.entity.UserMeeting;
 import com.example.meetingex.jwt.JwtUtil;
-import com.example.meetingex.repository.MeetingRepository;
-import com.example.meetingex.repository.UserMeetingRepository;
-import com.example.meetingex.repository.UserRepository;
+import com.example.meetingex.repository.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -24,7 +28,32 @@ public class MeetingService {
     private final UserRepository userRepository;
     private final MeetingRepository meetingRepository;
     private final UserMeetingRepository userMeetingRepository;
+    private final ScheduleRepository scheduleRepository;
+    private final UserScheduleRepository userScheduleRepository;
     private final JwtUtil jwtUtil;
+
+    public User findUserById(Long userId) {
+        return userRepository.findById(userId).get();
+    }
+
+    public boolean isMember(Long userID, Long meetingId) {
+        return userMeetingRepository.existsByUserIdAndMeetingId(userID, meetingId);
+    }
+
+    public void deleteUserScheduleByMeetingId(Long meetingId) {
+        //아이디를 찾고
+        List<Schedule> schedules = scheduleRepository.findByMeetingId(meetingId);
+        if (schedules == null)
+            return;
+
+        for(Schedule s : schedules) {
+            //그 아이디에 맞는 u-s를 자른다
+            userScheduleRepository.deleteByScheduleId(s.getId());
+            //그리고 스케쥴 본체도 자른다
+            scheduleRepository.delete(s);
+        }
+    }
+
 
     // 모임 생성
     @Transactional
@@ -63,10 +92,12 @@ public class MeetingService {
 
     //모임 수정
     //나중에 생성자만 수정 가능하도록
-
     @Transactional
-    public String updateMeeting(MeetingDto meetingDto, Long id) {
+    public String updateMeeting(MeetingDto meetingDto,String token, Long id) {
         try {
+            Long userId = jwtUtil.validateToken(token);
+            if (userId == null || !userMeetingRepository.findByUserIdAndMeetingId(userId, id).isCreator())
+                throw new Exception("유효하지 않은 유저임");
             if (meetingRepository.existsByName(meetingDto.getName()))
                 throw new Exception("중복임");
             Meeting meeting = meetingRepository.findById(id).get();
@@ -74,7 +105,6 @@ public class MeetingService {
             meeting.setDescription(meetingDto.getDescription());
             meeting.setMaxParticipants(meetingDto.getMaxParticipants());
             meetingRepository.save(meeting);
-            log.info("떴냐???@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@{}", meeting.getCurrentParticipants());
         } catch (Exception e) {
             return null;
         }
@@ -83,17 +113,26 @@ public class MeetingService {
 
     //모임 삭제, 일단 생성자는 배제시킨다
     @Transactional
-    public void deleteMeeting(Long id) {
+    public String deleteMeeting(Long id, String token) {
+        Long userId = jwtUtil.validateToken(token);
+        //유저 요청이 이상하거나 권한이 없다면 나가리
+        if (userId == null || !userMeetingRepository.findByUserIdAndMeetingId(userId, id).isCreator())
+            return null;
+
         //만약 스케쥴이 있다면?
         //스케쥴에 참여한 인원을 지우고
         //스케쥴을 지우고
+        deleteUserScheduleByMeetingId(id);
+
 
         //만약 참여자가 있다면?
         //참여자를 지우고
+        userMeetingRepository.deleteByMeetingId(id);
 
 
         //최종적으로 미팅을 지우게 만들면 될듯?
         meetingRepository.deleteById(id);
+        return "ok";
     }
 
     //모임 참여
@@ -102,6 +141,11 @@ public class MeetingService {
         Long userId = jwtUtil.validateToken(token);
         if (userId == null || !meetingRepository.existsById(id))
             return null;
+        System.out.println(userId);
+        //중복도 방지
+        if (userMeetingRepository.existsByUserIdAndMeetingId(userId, id))
+            return null;
+
         Meeting meeting = meetingRepository.findById(id).get();
         //꽉차도 나가리
         if (meeting.getCurrentParticipants() + 1 > meeting.getMaxParticipants())
@@ -115,6 +159,39 @@ public class MeetingService {
         userMeeting.setMeeting(meeting);
         userMeeting.setCreator(false);
         userMeetingRepository.save(userMeeting);
+        return "ok";
+    }
+
+    //참가자 조회
+    @Transactional(readOnly = true)
+    public List<ParticipantsDto> showParticipants(Long id) {
+        if (!meetingRepository.existsById(id))
+            return null;
+
+        return userMeetingRepository.findByMeetingId(id).stream()
+                .map(m -> new ParticipantsDto(m.getUser().getId(), m.getUser().getEmail()))
+                .collect(Collectors.toList());
+
+    }
+
+    //모임 탈퇴
+    @Transactional
+    public String exitMeeting(String token, Long meetingId) {
+        Long userId = jwtUtil.validateToken(token);
+        Meeting meeting = meetingRepository.findById(meetingId).get();
+        //아이디가 무효하거나 맴버가 아닌 사람이 나가거나 거기 안의 인원이 한명뿐일경우 나간다
+        if (userId == null || isMember(userId, meetingId))
+            return null;
+        UserMeeting userMeeting = userMeetingRepository.findByUserIdAndMeetingId(userId, meetingId);
+        //여기도 중복 부분 있어서 손봐야함
+
+        //암튼 지우는 주체가 생성자라면? -> 편의상 폭파
+        //1명뿐이라면 -> 편의상 폭파
+        if (userMeeting.isCreator() || meeting.getCurrentParticipants() == 1)
+            deleteMeeting(meetingId, token);
+        else {
+            userMeetingRepository.delete(userMeeting);
+        }
         return "ok";
     }
